- Refactored SerializeThrustcurveMotors.java to its own directory within thrustcurve, in preperation for benchmarks of serialization in next commit.
- Added user-input defined option for "% Of threads utilised" for developers to determine how many threads they want to dedicate to engine serialization.
