+++ core/src/main/java/module-info.java
@@ -100,8 +100,9 @@ open module info.openrocket.core {
 	exports info.openrocket.core.util.enums;
 	exports info.openrocket.core.utils;
 	exports info.openrocket.core.preferences;
+    exports info.openrocket.core.thrustcurve.serialization;
 
-	// Service providers
+    // Service providers
 	// Also edit core/src/main/resources/META-INF/services !! (until gradle-modules-plugin supports service
 	// copying, see https://github.com/java9-modularity/gradle-modules-plugin/issues/85)
 	provides info.openrocket.core.optimization.services.OptimizableParameterService with
